moved CredentialQueryResolverImpl to identity-hub-core
